+++ src/features/calculator/components/textField/HookedTextField.test.tsx
@@ -22,8 +22,10 @@ describe('Test text field component', () => {
       );
       const inputField = await screen.findByTestId('cartValue');
       await user.type(inputField, '123');
+      await user.tab();
       expect((inputField as HTMLInputElement).value).toBe('123');
     });
+
     test('Can input normal float number', async () => {
       const user = userEvent.setup();
       render(
@@ -37,8 +39,10 @@ describe('Test text field component', () => {
       );
       const inputField = await screen.findByTestId('cartValue');
       await user.type(inputField, '123.45');
+      await user.tab();
       expect((inputField as HTMLInputElement).value).toBe('123.45');
     });
+
     test('Can accept float number with plus sign', async () => {
       const user = userEvent.setup();
       render(
@@ -52,8 +56,52 @@ describe('Test text field component', () => {
       );
       const inputField = await screen.findByTestId('cartValue');
       await user.type(inputField, '+123.45');
+      await user.tab();
       expect((inputField as HTMLInputElement).value).toBe('123.45');
     });
+
+    test('Can not accept non-digit string', async () => {
+      const user = userEvent.setup();
+      render(
+        <Provider store={store}>
+          <HookedTextField
+            name="cartValue"
+            type="float"
+            label="Value in Euros"
+          />
+        </Provider>,
+      );
+      const inputField = await screen.findByTestId('cartValue');
+      await user.type(inputField, 'test');
+      await user.tab();
+      expect((inputField as HTMLInputElement).value).toBe('test');
+      const errorMessage = await screen.findByText(
+        'Value must be a float number',
+      );
+      expect(errorMessage).toBeDefined();
+    });
+
+    test('Can not accept plus sign with random input', async () => {
+      const user = userEvent.setup();
+      render(
+        <Provider store={store}>
+          <HookedTextField
+            name="cartValue"
+            type="float"
+            label="Value in Euros"
+          />
+        </Provider>,
+      );
+      const inputField = await screen.findByTestId('cartValue');
+      await user.type(inputField, '+123-1');
+      await user.tab();
+      expect((inputField as HTMLInputElement).value).toBe('+123-1');
+      const errorMessage = await screen.findByText(
+        'Value must be a float number',
+      );
+      expect(errorMessage).toBeDefined();
+    });
+
     test('Can not accept negative number', async () => {
       const user = userEvent.setup();
       render(
@@ -67,12 +115,14 @@ describe('Test text field component', () => {
       );
       const inputField = await screen.findByTestId('cartValue');
       await user.type(inputField, '-');
-      expect((inputField as HTMLInputElement).value).toBe('');
+      await user.tab();
+      expect((inputField as HTMLInputElement).value).toBe('-');
       const errorMessage = await screen.findByText(
         'Value must be a positive float number',
       );
       expect(errorMessage).toBeDefined();
     });
+
     test('Can not accept non-digit character', async () => {
       const user = userEvent.setup();
       render(
@@ -86,7 +136,8 @@ describe('Test text field component', () => {
       );
       const inputField = await screen.findByTestId('cartValue');
       await user.type(inputField, '123.+++');
-      expect((inputField as HTMLInputElement).value).toBe('123.');
+      await user.tab();
+      expect((inputField as HTMLInputElement).value).toBe('123.+++');
       const errorMessage = await screen.findByText(
         'Value must be a float number',
       );
@@ -108,8 +159,10 @@ describe('Test text field component', () => {
       );
       const inputField = await screen.findByTestId('cartValue');
       await user.type(inputField, '123');
+      await user.tab();
       expect((inputField as HTMLInputElement).value).toBe('123');
     });
+
     test('Can not accept float number', async () => {
       const user = userEvent.setup();
       render(
@@ -123,10 +176,12 @@ describe('Test text field component', () => {
       );
       const inputField = await screen.findByTestId('cartValue');
       await user.type(inputField, '123.');
-      expect((inputField as HTMLInputElement).value).toBe('123');
+      await user.tab();
+      expect((inputField as HTMLInputElement).value).toBe('123.');
       const errorMessage = await screen.findByText('Value must be an integer');
       expect(errorMessage).toBeDefined();
     });
+
     test('Can accept integer with plus sign', async () => {
       const user = userEvent.setup();
       render(
@@ -140,8 +195,48 @@ describe('Test text field component', () => {
       );
       const inputField = await screen.findByTestId('cartValue');
       await user.type(inputField, '+123');
+      await user.tab();
       expect((inputField as HTMLInputElement).value).toBe('123');
     });
+
+    test('Can not accept plus sign with random input', async () => {
+      const user = userEvent.setup();
+      render(
+        <Provider store={store}>
+          <HookedTextField
+            name="cartValue"
+            type="number"
+            label="Value in Euros"
+          />
+        </Provider>,
+      );
+      const inputField = await screen.findByTestId('cartValue');
+      await user.type(inputField, '+123+-n');
+      await user.tab();
+      expect((inputField as HTMLInputElement).value).toBe('+123+-n');
+      const errorMessage = await screen.findByText('Value must be an integer');
+      expect(errorMessage).toBeDefined();
+    });
+
+    test('Can not accept non-digit string', async () => {
+      const user = userEvent.setup();
+      render(
+        <Provider store={store}>
+          <HookedTextField
+            name="cartValue"
+            type="number"
+            label="Value in Euros"
+          />
+        </Provider>,
+      );
+      const inputField = await screen.findByTestId('cartValue');
+      await user.type(inputField, 'test');
+      await user.tab();
+      expect((inputField as HTMLInputElement).value).toBe('test');
+      const errorMessage = await screen.findByText('Value must be an integer');
+      expect(errorMessage).toBeDefined();
+    });
+
     test('Can not accept negative number', async () => {
       const user = userEvent.setup();
       render(
@@ -155,12 +250,14 @@ describe('Test text field component', () => {
       );
       const inputField = await screen.findByTestId('cartValue');
       await user.type(inputField, '-');
-      expect((inputField as HTMLInputElement).value).toBe('');
+      await user.tab();
+      expect((inputField as HTMLInputElement).value).toBe('-');
       const errorMessage = await screen.findByText(
         'Value must be a positive integer',
       );
       expect(errorMessage).toBeDefined();
     });
+
     test('Can not accept non-digit character', async () => {
       const user = userEvent.setup();
       render(
@@ -174,7 +271,8 @@ describe('Test text field component', () => {
       );
       const inputField = await screen.findByTestId('cartValue');
       await user.type(inputField, '123+.++');
-      expect((inputField as HTMLInputElement).value).toBe('123');
+      await user.tab();
+      expect((inputField as HTMLInputElement).value).toBe('123+.++');
       const errorMessage = await screen.findByText('Value must be an integer');
       expect(errorMessage).toBeDefined();
     });