+++ CHANGELOG.md
@@ -5,12 +5,21 @@
 - **Currently, all links to the commits do not work as the project is hold on GitHub as a PRIVATE repo. I am happy to provide access for review upon request.**
 
 
+## v0.1.3
+### Features
+- [`bc3349f`](https://github.com/wangc9/delivery-fee-calculator/commit/bc3349f4194b47850a9fdc1a4ea4f0eca29f14ec)* Change input validation
+	- Validation now happens only when the user has finished typing and clicked out of the text box
+	- The submit button is disabled until all text boxes receive correct input (positive integer or float number and order time is later than the current time)
+	- All unit and integration have been updated
+	- e2e tests include edge cases
+
+
 ## v0.1.2
 ### Features
-- [`1fc5a23`](https://github.com/wangc9/delivery-fee-calculator/commit/1fc5a23a1403cbf7ae89d553b8bf46cff55383fb)* Change front page style
+- [`bc3349f`](https://github.com/wangc9/delivery-fee-calculator/commit/bc3349f4194b47850a9fdc1a4ea4f0eca29f14ec)* Change front page style
 
 ### Chores
-- [`1fc5a23`](https://github.com/wangc9/delivery-fee-calculator/commit/1fc5a23a1403cbf7ae89d553b8bf46cff55383fb)* Add README for local build instructions
+- [`bc3349f`](https://github.com/wangc9/delivery-fee-calculator/commit/bc3349f4194b47850a9fdc1a4ea4f0eca29f14ec)* Add README for local build instructions
 
 
 ## v0.1.1